LayoutTests/canvas/philip/tests/toDataURL.jpeg.alpha.html

<!DOCTYPE html>
<title>Canvas test: toDataURL.jpeg.alpha</title>
<script src="../tests.js"></script>
<link rel="stylesheet" href="../tests.css">
<body>
<p id="passtext">Pass</p>
<p id="failtext">Fail</p>
<p class="output">These images should be identical:</p>
<canvas id="c" class="output" width="100" height="50"><p class="fallback">FAIL (fallback content)</p></canvas>
<p class="output expectedtext">Expected output:<p><img src="toDataURL.jpeg.alpha.png" class="output expected" id="expected" alt="">
<ul id="d"></ul>
<script>
_addTest(function(canvas, ctx) {

ctx.fillStyle = 'rgba(128, 255, 128, 0.5)';
ctx.fillRect(0, 0, 100, 50);
ctx.globalCompositeOperation = 'destination-over'; // should be ignored by toDataURL
var data = canvas.toDataURL('image/jpeg');
ctx.globalCompositeOperation = 'source-over';
if (!data.match(/^data:image\/jpeg[;,]/)) {
  _assert(true, "true");
} else {
  ctx.fillStyle = '#f00';
  ctx.fillRect(0, 0, 100, 50);
  var img = new Image();
  deferTest();
  img.onload = wrapFunction(function ()
  {
      ctx.drawImage(img, 0, 0);
      
      _assertPixelApprox(canvas, 50,25, 63,127,63,255, "50,25", "63,127,63,255", 8);
  });
  img.src = data;
}


});
</script>
