LayoutTests/canvas/philip/tests/toDataURL.jpeg.alpha-expected.txt

< [index] >
toDataURL.jpeg.alpha
toDataURL with JPEG composites onto black
References: toDataURL.jpeg toDataURL.noalpha
Actual output:
Expected output:

Passed
